fixing name in model and giving confident in bottomsheet

+++ app/src/main/java/academy/bangkit/capstonk/foories/core/data/source/remote/entity/FoodPayload.kt
@@ -2,5 +2,5 @@ package academy.bangkit.capstonk.foories.core.data.source.remote.entity
 
 data class FoodPayload(
     val name: String?,
-    val confidence: Float?
+    val confidence: Double?
 )
